Filled in Carlot.activateEvents; began working with Carlot.cardFocus
Carlot.activateEvents now adds event listeners to each card with an anonymous callback function that, when clicked, enables the text input and gives it focus and passes the unique id of the card clicked to Carlot.cardFocus. For now, Carlot.cardFocus just toggles the class "card-focus" on the card clicked.